S3C-35 Refactoring of bucketfile, V2 (S3 file metadata backend)
The changes allow bucketfile to use a new API in Arsenal to
communicate with a remote leveldb database, containing sublevels for
bucket storage. Metadata is still stored on a local levelDB server,
but it should now be easy to move the storage logic in a new daemon
running on a remote server, and it should be robust.

Instead of relying on the existing implementation of multilevel, it
uses client/server wrappers around a new level-net communication
protocol and API in Arsenal based on socket.io to exchange messages.

It shall be compatible with the existing metadata since it still uses
the core sublevel module for the storage logic, only the RPC procotol
has changed.

Test done:

 - put a few 100s of files in different S3 subdirectories
 - list directories and subdirectories
 - get/delete files
 - multi-part upload
 - introduce random connection errors (tcpkill) to check robustness
   and automatic reconnection